requirements.txt
altair==4.2.0
millify==0.1.1
pandas==1.5.2
Pillow==9.4.0
plotly==5.11.0
python-dotenv==0.21.0
requests==2.28.1
streamlit==1.16.0
streamlit_option_menu==0.3.2
